Fixed FASTA missing ">name" as first line error
When a FASTA input file without a first "name line" was supplied, Biopython.SeqIO does not iterate, yielding an error in line 610 (i2fasta[i] -> KeyError: 0) in intronserter.py. This is because the complete file is not parsed by Biopython. No error message is given by Biopython. Therefore, a file manipulation was introduced as the very first step: if the FASTA file does not contain a first line starting with ">", this line is added to the input FASTA file.